fixed cell text overwriting cell border
by adding a subpad to each cell pad that sits at (1,1) with y-2 and x-2 of
the original pad (thus fitting in the pad) we can write the text to the
subpad instead and dont have to worry about overwriting the border.

Also needed to switch from compiling with lncurses to lncursesw and
changing the local in main.c to handle umlauts.